dcompact_worker: add conf param oneshot_dbname_prefix
when `copy_sst_files` is true, dcompact will create dir `hoster_root/dbname/job-%5d`,
the dir `hoster_root/dbname` needs to be deleted if it is a oneshot db, such as
`.rocksdb/default-XXXXXX` in MyTopling.

We add conf param `oneshot_dbname_prefix` to recognize such dir to be deleted.