dcompact_worker: AutoStartZipServer must be strong symbol
`AutoStartZipServer` is in plugin `ToplingZipTable`.

In open source community version, it just print a message, there
is no ZipServer.

In enterprise version, ZipServer will be started.

Previously, we declare AutoStartZipServer as weak symbol, thus
when linked with community 'librocksdb.so', AutoStartZipServer
is not defined, when community 'librocksdb.so' is replaced by
enterprise 'librocksdb.so', AutoStartZipServer exists, but the
weak symbol is still NULL, unless 'librocksdb.so' is added to
LD_PRELOAD, the loader will not resolve AutoStartZipServer.
This behavior is strange, this commit will prevent the strange
error from happening.